+++ src/bin/tikv-server.rs
@@ -29,6 +29,8 @@ extern crate signal;
 extern crate nix;
 extern crate prometheus;
 
+mod signal_handler;
+
 use std::process;
 use std::{env, thread};
 use std::fs::{self, File};
@@ -43,7 +45,6 @@ use getopts::{Options, Matches};
 use rocksdb::{DB, Options as RocksdbOptions, BlockBasedOptions};
 use mio::EventLoop;
 use fs2::FileExt;
-use prometheus::{Encoder, TextEncoder};
 
 use tikv::storage::{Storage, TEMP_DIR, ALL_CFS};
 use tikv::util::{self, panic_hook, rocksdb as rocksdb_util};
@@ -58,9 +59,6 @@ use tikv::server::{PdStoreAddrResolver, StoreAddrResolver};
 use tikv::raftstore::store::{self, SnapManager};
 use tikv::pd::RpcClient;
 use tikv::util::time_monitor::TimeMonitor;
-
-const ROCKSDB_DB_STATS_KEY: &'static str = "rocksdb.dbstats";
-const ROCKSDB_CF_STATS_KEY: &'static str = "rocksdb.cfstats";
 
 fn print_usage(program: &str, opts: Options) {
     let brief = format!("Usage: {} [options]", program);
@@ -613,72 +611,14 @@ fn start_server<T, S>(mut server: Server<T, S>,
 {
     let ch = server.get_sendch();
     let h = thread::Builder::new()
-        .name("tikv-server".to_owned())
+        .name("server".to_owned())
         .spawn(move || {
             server.run(&mut el).unwrap();
         })
         .unwrap();
-    handle_signal(ch, engine, backup_path);
+    signal_handler::handle_signal(ch, engine, backup_path);
     h.join().unwrap();
 }
-
-#[cfg(unix)]
-fn handle_signal(ch: SendCh<Msg>, engine: Arc<DB>, backup_path: &str) {
-    use signal::trap::Trap;
-    use nix::sys::signal::{SIGTERM, SIGINT, SIGUSR1, SIGUSR2};
-    let trap = Trap::trap(&[SIGTERM, SIGINT, SIGUSR1, SIGUSR2]);
-    for sig in trap {
-        match sig {
-            SIGTERM | SIGINT => {
-                info!("receive signal {}, stopping server...", sig);
-                ch.send(Msg::Quit).unwrap();
-                break;
-            }
-            SIGUSR1 => {
-                // Use SIGUSR1 to log metrics.
-                let mut buffer = vec![];
-                let metric_familys = prometheus::gather();
-                let encoder = TextEncoder::new();
-                encoder.encode(&metric_familys, &mut buffer).unwrap();
-                info!("{}", String::from_utf8(buffer).unwrap());
-
-                // Log common rocksdb stats.
-                for name in engine.cf_names() {
-                    let handler = engine.cf_handle(name).unwrap();
-                    if let Some(v) = engine.get_property_value_cf(handler, ROCKSDB_CF_STATS_KEY) {
-                        info!("{}", v)
-                    }
-                }
-
-                if let Some(v) = engine.get_property_value(ROCKSDB_DB_STATS_KEY) {
-                    info!("{}", v)
-                }
-
-                // Log more stats if enable_statistics is true.
-                if let Some(v) = engine.get_statistics() {
-                    info!("{}", v)
-                }
-            }
-            SIGUSR2 => {
-                if backup_path.is_empty() {
-                    info!("empty backup path, backup is disabled");
-                    continue;
-                }
-
-                info!("backup db to {}", backup_path);
-                if let Err(e) = engine.backup_at(backup_path) {
-                    error!("fail to backup: {}", e);
-                }
-                info!("backup done");
-            }
-            // TODO: handle more signal
-            _ => unreachable!(),
-        }
-    }
-}
-
-#[cfg(not(unix))]
-fn handle_signal(_: SendCh<Msg>, _: Arc<DB>, _: &str) {}
 
 fn run_raft_server(pd_client: RpcClient, cfg: Config, backup_path: &str, config: &toml::Value) {
     let mut event_loop = create_event_loop(&cfg).unwrap();
